create_directory: work around mkdirat/renameat race to avoid ENOENT
Very interesting race condition when copyup'ing directories.

If multiple processes (or nodes) perform a copyup of the same path
at the same time, create_directory() ends with a renameat() that is
used for each directory to enforce atomicity. However, IF the
directory is still empty, another renameat() will also succeed and
override the empty directory with a new one.  IF this happens between
openat and mkdirat of another thread's create_directory(), mkdirat
will fail with ENOENT.

The most elegant way to fix this would be to use renameat2() but Lustre
doesn't implement it yet so we can't avoid this behavior at this time.

This change implements a retry for mkdirat by reopening the directory.