rpl_stat: introduce st_nlink caching for multilayer directories (node->dircnt)
rpl_stat() being called by ovl_lookup(), it seems to be called very
often on the current working directory (cwd). For very large directories,
the cost of iterating over all children to compute the st_nlink value can
thus be very important.

So this patch should help when the cwd is a multilayer directory containing
a lot of entries.