ovl_do_open: retain node on file create to avoid race condition
This patch actually fixes two issues:

1. when a file is created, this is done by ovl_create -> ovl_do_open,
with a lookup (that retains it) done afterwards, leading to a possible
race condition if node_free() is called from another thread in between
ovl_do_open() and do_lookup_file().

ovl_do_open() can now retain the node so that it can't be forgotten
until the lookup is performed in ovl_create().

2. there was a possible local race condition in ovl_do_open() between
openat() and make_ovl_node/insert_node in replace mode. We now take a
lock on the parent directory to ensure the node is not created by a
parallel load_dir() occurring between openat() and insert_node().

Thanks to Andrew Beel for the report and reproducer.